test(server): test that room name and opening status are persisted into database
